move pytrigno library to separate ros2 package
