+++ pytrigno/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script-dir=$base/lib/pytrigno
+[install]
+install-scripts=$base/lib/pytrigno
